requirements.txt
boto3==1.42.96
category_encoders==2.9.0
fastapi[standard]==0.136.1
joblib==1.5.3
numpy==2.4.4
pandas==3.0.2
psycopg2_binary==2.9.11
pydantic==2.13.3
python-dotenv==1.2.2
scikit_learn==1.8.0